requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
mcp>=1.0.0
requests>=2.31.0
colorama>=0.4.6
